+++ vibra/engine/analysis_info/modal_analysis_setup.py
@@ -1,23 +1,7 @@
-from dataclasses import dataclass, fields, KW_ONLY 
-from functools import wraps
+from dataclasses import KW_ONLY, dataclass
+
 from vibra.engine.analysis_info import AnalysisID
-
-def ignore_extra_kwargs(cls):
-    original_init = cls.__init__
-
-    @wraps(original_init)
-    def new_init(self, *args, **kwargs):
-        
-        # expected fields of original dataclass
-        expected_fields = {f.name for f in fields(cls)}
-
-        # filter the unnecessary kwargs
-        filtered_kwargs = {k: v for k, v in kwargs.items() if k in expected_fields}
-
-        original_init(self, *args, **filtered_kwargs)
-
-    cls.__init__ = new_init
-    return cls
+from vibra.utils.dataclass_utils import ignore_extra_kwargs
 
 
 @ignore_extra_kwargs
